portfolio-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0

